Delete onBoardingActions and rename Chat and Conversation Screens
