+++ src/actions/chatActions.js
@@ -12,8 +12,10 @@ const SEND_MESSAGE = 'SEND_MESSAGE';
 const SUBSCRIBE = 'SUBSCRIBE';
 const UNSUBSCRIBE = 'UNSUBSCRIBE';
 
+export const CLEAR_CHAT_STATE = 'CLEAR_CHAT_STATE';
 export const RECEIVE_MESSAGE = 'RECEIVE_MESSAGE';
 
+export const clearChatState = createAction(CLEAR_CHAT_STATE);
 export const createConsumer = createAction(CREATE_CONSUMER);
 export const disconnectActionCable = createAction(DISCONNECT_ACTION_CABLE);
 export const receiveMessage = createAction(RECEIVE_MESSAGE);
@@ -30,7 +32,7 @@ export const getConversations = createThunk(GET_CONVERSATIONS, async () => {
   }
 });
 
-// get all messages from a conversacion (1 conversation = * messages)
+// get all messages from a conversation (1 conversation = * messages)
 export const getMessages = createThunk(GET_MESSAGES, async (id, page) => {
   try {
     return (await chatService.getMessages(id, page)).data.messages;
