--- src/actions/onboardingActions.js
@@ -1,5 +1,0 @@
-import { createAction } from '@rootstrap/redux-tools';
-
-const CLEAN_ONBOARDING = 'CLEAN_ONBOARDING';
-
-export const clearOnboarding = createAction(CLEAN_ONBOARDING);
