+++ src/constants/screens.js
@@ -1,6 +1,6 @@
 export const CAMERA_ROLL_SCREEN = 'CameraRollScreen';
+export const CHAT_LIST_SCREEN = 'ChatListScreen';
 export const CHAT_SCREEN = 'ChatScreen';
-export const CONVERSATION_SCREEN = 'ConversationScreen';
 export const LOGIN_SCREEN = 'LoginScreen';
 export const MAIN_SCREEN = 'MainScreen';
 export const PROFILE_SCREEN = 'ProfileScreen';
