+++ src/navigators/index.js
@@ -9,12 +9,12 @@ import mark from 'assets/images/mark.png';
 import profileIcon from 'assets/images/profile.png';
 import NavHeader from 'components/common/NavHeader';
 import { targetMatchParam } from 'constants/parameters';
-import { CHAT_SCREEN, PROFILE_SCREEN } from 'constants/screens';
+import { CHAT_LIST_SCREEN, PROFILE_SCREEN } from 'constants/screens';
 import strings from 'locale';
 import AppLoader from 'screens/AppLoader';
 import CameraRollScreen from 'screens/CameraRollScreen';
+import ChatListScreen from 'screens/ChatListScreen';
 import ChatScreen from 'screens/ChatScreen';
-import ConversationScreen from 'screens/ConversationScreen';
 import LoginScreen from 'screens/LoginScreen';
 import MainScreen from 'screens/MainScreen';
 import ProfileScreen from 'screens/ProfileScreen';
@@ -58,7 +58,7 @@ const MainNavigator = createStackNavigator({
       headerRight: (
         <NavHeader
           icon={chatBubble}
-          onPress={() => navigation.push(CHAT_SCREEN)}
+          onPress={() => navigation.push(CHAT_LIST_SCREEN)}
           style={styles.rightIcon}
         />
       ),
@@ -92,8 +92,8 @@ const MainNavigator = createStackNavigator({
     }),
   },
 
-  ChatScreen: {
-    screen: ChatScreen,
+  ChatListScreen: {
+    screen: ChatListScreen,
     navigationOptions: ({ navigation }) => ({
       headerLeft: (
         <NavHeader
@@ -113,8 +113,8 @@ const MainNavigator = createStackNavigator({
     }),
   },
 
-  ConversationScreen: {
-    screen: ConversationScreen,
+  ChatScreen: {
+    screen: ChatScreen,
     navigationOptions: ({
       navigation,
       navigation: {
